# docs/website/openapi/websocket.yaml
openapi: 3.0.0
info:
  title: Owlistic API
  description: API documentation for Owlistic application.
  version: 1.0.0
servers:
  - url: http://localhost:8080/api/v1
paths:
...
  /blocks:
    get:
      summary: Get blocks
      responses:
        '200':
          description: List of blocks
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  properties:
                    id:
                      type: string
                    note_id:
                      type: string
                    type:
                      type: string
        '401':
          description: User not authenticated
        '500':
          description: Failed to retrieve blocks

    post:
      summary: Create a new block
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                content:
                  type: object
                note_id:
                  type: string
              required:
                - content
                - note_id
      responses:
        '201':
          description: Block created successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: string
        '400':
          description: Invalid input
        '401':
          description: User not authenticated
        '500':
          description: Failed to create block
  /blocks/{id}:
    get:
      summary: Get block by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Block found
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: string
                  note_id:
                    type: string
                  type:
                    type: string
        '404':
          description: Block not found
        '401':
          description: User not authenticated
        '500':
          description: Failed to retrieve block

    put:
      summary: Update a block
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                content:
                  type: object
              required:
                - content
      responses:
        '200':
          description: Block updated successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: string
                  note_id:
                    type: string
                  type:
                    type: string
        '404':
          description: Block not found
        '401':
          description: User not authenticated
        '500':
          description: Failed to update block

    delete:
      summary: Delete a block
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '204':
          description: Block deleted successfully
        '404':
          description: Block not found
        '401':
          description: User not authenticated
        '500':
          description: Failed to delete block